Regroup the HLS tests by production file
Each HLS test file now covers one production file. Pure moves plus the
router rewrite; no assertion changed.

- hls_actual_start_test.go is gone. Its three writeHLSPlaylistHeaders
  tests belong to hls_handler.go and moved to hls_handler_test.go; its two
  measureHLSSessionStart tests moved to hls_session_test.go.
- hls_session_cache_test.go is new: the reservation, per-user cap, LRU
  eviction, idle reclaim, TTL refresh and owner cleanup tests, which were
  split across hls_session_test.go and hls_handler_test.go. These pin
  lock ordering and "never kill another device's active stream", so they
  moved verbatim.
- TestHLSSessionKey moved from hls_handler_test.go to hls_session_test.go.
- The five TestBuildHLSPlaylistBody* functions became subtests of one
  TestBuildHLSPlaylistBody, each keeping its audit comment. One of them
  had drifted 700 lines away from the other four.
- The six tests that hand-built a chi router with session middleware now
  call newHLSTestHandler.

Coverage holds at 69.7%; per-function coverage is byte-identical.
